working my way through the bugs
